Remove unused import
